refactor: Extract logs menu to LogsMenuMixin
- Create logs_menu_mixin.py with _logs_menu and _view_meshforge_logs
- Reduces main.py from 1,548 to 1,431 lines (now under 1,500 target)
- Continues file size reduction effort per CLAUDE.md guidelines
